+++ src/test/java/Zoo2Test.java
@@ -1,5 +1,4 @@
 import animal.Animal;
-import animal.AnimalType;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -12,7 +11,7 @@ class Zoo2Test {
 
     @Test
     void addAnimal() {
-        var zoo = new StrategicalZoo();
+        var zoo = new NamedZoo(new StrategicalZoo());
         assertEquals(0, zoo.getAnimalCount(), "no animal inside");
         zoo.addAnimal(WOLF.create("Petya"));
         assertEquals(1, zoo.getAnimalCount(), "one animal inside");
@@ -20,7 +19,7 @@ class Zoo2Test {
 
     @Test
     void removeAnimal() {
-        var zoo = new StrategicalZoo();
+        var zoo = new NamedZoo(new StrategicalZoo());
         assertEquals(0, zoo.getAnimalCount(), "no animal inside");
         var animal = WOLF.create("Vasya");
 
@@ -33,7 +32,7 @@ class Zoo2Test {
 
     @Test
     void getAnimalByNumber() {
-        var zoo = new StrategicalZoo();
+        var zoo = new NamedZoo(new StrategicalZoo());
         assertEquals(0, zoo.getAnimalCount(), "no animal inside");
 
         zoo.addAnimal(WOLF.create("Petya"));
@@ -51,7 +50,7 @@ class Zoo2Test {
 
     @Test
     void getAnimalsByNames() {
-        var zoo = new StrategicalZoo();
+        var zoo = new NamedZoo(new StrategicalZoo());
         assertEquals(0, zoo.getAnimalCount(), "no animal inside");
 
         var petyaNameString = "Petya";
